removed useless files, made essentials files for fetching data
