+++ client/app/components/layout/banner/Banner.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import './banner.scss'
 
 export default function Banner() {
   return (
