--- client/app/components/layout/banner/banner.scss
@@ -1,44 +1,0 @@
-.banner-wrapper {
-  background: url('/background.png') no-repeat bottom center;
-  background-size: cover;
-  height: 300px;
-  width: 100%;
-
-  @media only screen and (max-width: 768px) {
-    height: 200px;
-  }
-
-.banner-background {
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.4);
-}
-
-  .banner-main {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    height: 100%;
-    color: white;
-    padding-left: 40px;
-
-    @media only screen and (max-width: 768px) {
-      width: 100%;
-      align-items: center;
-      justify-content: center;
-      padding-left: 0;
-    }
-
-    .banner-title {
-      font-size: 1.8rem;
-      font-weight: 700;
-      margin-bottom: 0;
-
-      .banner-description {
-        @media only screen and (max-width: 768px) {
-          font-size: 1.2rem;
-        }
-      }
-    }
-  }
-}
