+++ client/app/components/layout/footer/Footer.tsx
@@ -2,8 +2,6 @@ import React from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
 
-import './footer.scss'
-
 export default function Footer() {
   return (
     <div className='footer-wrapper'>
