--- client/app/components/layout/nav/nav.scss
@@ -1,76 +1,0 @@
-@use '../../../styles/globals.scss';
-
-nav {
-  background-color: globals.$primary-color;
-  color: white;
-  display: flex;
-  justify-content: space-between;
-  position: sticky;
-  top: 0;
-  width: 100%;
-  height: 50px;
-  z-index: 1000;
-
-  .nav-logo-container {
-    align-items: center;
-    display: flex;
-    height: 100%;
-    justify-content: center;
-    padding-left: 20px;
-    text-decoration: none;
-    font-size: 1.5rem;
-    font-weight: 700;
-
-    a{
-      color: white;
-      text-decoration: none;
-
-      .logo-green {
-        color: #7acc88;
-      }
-    }
-  }
-
-  .nav-links-container {
-    align-items: center;
-    display: flex;
-    height: 100%;
-    justify-content: center;
-    padding-right: 20px;
-    text-decoration: none;
-    font-size: 1.1rem;
-
-    .nav-user-data {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      .nav-user-data-username {
-        font-size: 1.1rem;
-        margin-right: 5px;
-      }
-    }
-
-    a {
-      color: white;
-      margin-left: 20px;
-      text-decoration: none;
-
-      .profile-button {
-        width: 32px;
-        transform: scale(1);
-        transition: transform 0.3s ease;
-      }
-
-      .profile-button:hover {
-        transform: scale(1.1);
-        transition: transform 0.3s ease;
-      }
-    }
-
-    a:hover {
-      color: #fccad6;
-      transition: color 0.3s ease;
-    }
-  }
-}
