+++ client/app/essentials/types.tsx
@@ -1,0 +1,4 @@
+export interface UserData {
+  id: string;
+  username: string;
+}
