+++ client/app/styles/layout.scss
@@ -1,4 +1,124 @@
-@use '../../../styles/globals.scss';
+@use './globals.scss';
+
+nav {
+  background-color: globals.$primary-color;
+  color: white;
+  display: flex;
+  justify-content: space-between;
+  position: sticky;
+  top: 0;
+  width: 100%;
+  height: 50px;
+  z-index: 1000;
+
+  .nav-logo-container {
+    align-items: center;
+    display: flex;
+    height: 100%;
+    justify-content: center;
+    padding-left: 20px;
+    text-decoration: none;
+    font-size: 1.5rem;
+    font-weight: 700;
+
+    a{
+      color: white;
+      text-decoration: none;
+
+      .logo-green {
+        color: #7acc88;
+      }
+    }
+  }
+
+  .nav-links-container {
+    align-items: center;
+    display: flex;
+    height: 100%;
+    justify-content: center;
+    padding-right: 20px;
+    text-decoration: none;
+    font-size: 1.1rem;
+
+    .nav-user-data {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+
+      .nav-user-data-username {
+        font-size: 1.1rem;
+        margin-right: 5px;
+      }
+    }
+
+    a {
+      color: white;
+      margin-left: 20px;
+      text-decoration: none;
+
+      .profile-button {
+        width: 32px;
+        transform: scale(1);
+        transition: transform 0.3s ease;
+      }
+
+      .profile-button:hover {
+        transform: scale(1.1);
+        transition: transform 0.3s ease;
+      }
+    }
+
+    a:hover {
+      color: #fccad6;
+      transition: color 0.3s ease;
+    }
+  }
+}
+
+.banner-wrapper {
+  background: url('/background.png') no-repeat bottom center;
+  background-size: cover;
+  height: 300px;
+  width: 100%;
+
+  @media only screen and (max-width: 768px) {
+    height: 200px;
+  }
+
+.banner-background {
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.4);
+}
+
+  .banner-main {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    height: 100%;
+    color: white;
+    padding-left: 40px;
+
+    @media only screen and (max-width: 768px) {
+      width: 100%;
+      align-items: center;
+      justify-content: center;
+      padding-left: 0;
+    }
+
+    .banner-title {
+      font-size: 1.8rem;
+      font-weight: 700;
+      margin-bottom: 0;
+
+      .banner-description {
+        @media only screen and (max-width: 768px) {
+          font-size: 1.2rem;
+        }
+      }
+    }
+  }
+}
 
 .footer-wrapper {
   display: flex;
